requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase>=2.4.0
flask-limiter>=4.0.0
PyJWT>=2.0.0
Werkzeug>=3.0.0
bcrypt>=4.0.0